first code seperation, no locks yet
